Move internal signing interfaces into Internal folder and namespace.
