+++ src/Altinn.App.Core/Features/Signing/Interfaces/Internal/ISignDocumentManager.cs
@@ -2,7 +2,7 @@ using Altinn.App.Core.Features.Signing.Models.Internal;
 using Altinn.App.Core.Internal.Process.Elements.AltinnExtensionProperties;
 using Altinn.Platform.Storage.Interface.Models;
 
-namespace Altinn.App.Core.Features.Signing.Interfaces;
+namespace Altinn.App.Core.Features.Signing.Interfaces.Internal;
 
 internal interface ISignDocumentManager
 {
