+++ src/Altinn.App.Core/Features/Signing/Interfaces/Internal/ISigneeContextsManager.cs
@@ -1,7 +1,7 @@
 using Altinn.App.Core.Features.Signing.Models.Internal;
 using Altinn.App.Core.Internal.Process.Elements.AltinnExtensionProperties;
 
-namespace Altinn.App.Core.Features.Signing.Interfaces;
+namespace Altinn.App.Core.Features.Signing.Interfaces.Internal;
 
 internal interface ISigneeContextsManager
 {
