+++ src/Altinn.App.Core/Features/Signing/Interfaces/Internal/ISigningCallToActionService.cs
@@ -4,7 +4,7 @@ using Altinn.App.Core.Internal.Process.Elements.AltinnExtensionProperties;
 using Altinn.App.Core.Models;
 using Altinn.Platform.Register.Models;
 
-namespace Altinn.App.Core.Features.Signing.Interfaces;
+namespace Altinn.App.Core.Features.Signing.Interfaces.Internal;
 
 /// <summary>
 /// Interface for sending correspondence to alert a signee of a signing call to action.
