+++ src/Altinn.App.Core/Features/Signing/Interfaces/Internal/ISigningService.cs
@@ -2,7 +2,7 @@
 using Altinn.App.Core.Internal.Process.Elements.AltinnExtensionProperties;
 using static Altinn.App.Core.Features.Signing.Models.Internal.Signee;
 
-namespace Altinn.App.Core.Features.Signing.Interfaces;
+namespace Altinn.App.Core.Features.Signing.Interfaces.Internal;
 
 internal interface ISigningService
 {
