+++ src/Altinn.App.Core/Features/Validation/Default/SigningTaskValidator.cs
@@ -1,5 +1,5 @@
 using System.Diagnostics;
-using Altinn.App.Core.Features.Signing.Interfaces;
+using Altinn.App.Core.Features.Signing.Interfaces.Internal;
 using Altinn.App.Core.Internal.App;
 using Altinn.App.Core.Internal.Process;
 using Altinn.App.Core.Internal.Process.Elements.AltinnExtensionProperties;
