+++ src/Altinn.App.Core/Internal/Process/ProcessTasks/SigningProcessTask.cs
@@ -1,5 +1,5 @@
 using Altinn.App.Core.Features;
-using Altinn.App.Core.Features.Signing.Interfaces;
+using Altinn.App.Core.Features.Signing.Interfaces.Internal;
 using Altinn.App.Core.Features.Signing.Models.Internal;
 using Altinn.App.Core.Helpers;
 using Altinn.App.Core.Internal.App;
