+++ test/Altinn.App.Api.Tests/Controllers/SigningControllerTests.cs
@@ -3,7 +3,7 @@ using Altinn.App.Api.Controllers;
 using Altinn.App.Api.Models;
 using Altinn.App.Core.Configuration;
 using Altinn.App.Core.Features.Auth;
-using Altinn.App.Core.Features.Signing.Interfaces;
+using Altinn.App.Core.Features.Signing.Interfaces.Internal;
 using Altinn.App.Core.Features.Signing.Models.Internal;
 using Altinn.App.Core.Helpers.Serialization;
 using Altinn.App.Core.Internal.App;
