+++ test/Altinn.App.Core.Tests/Internal/Process/ProcessTasks/SigningProcessTaskTests.cs
@@ -1,6 +1,6 @@
 using Altinn.App.Core.Configuration;
 using Altinn.App.Core.Features;
-using Altinn.App.Core.Features.Signing.Interfaces;
+using Altinn.App.Core.Features.Signing.Interfaces.Internal;
 using Altinn.App.Core.Features.Signing.Models.Internal;
 using Altinn.App.Core.Helpers.Serialization;
 using Altinn.App.Core.Internal.App;
